// src/interfaces/wms/transaction/outbound/outboundJobWms.interface.ts
export interface IOutboundJobView {
  company_code: string;
  prin_code: string;
  prin_name: string;
  job_class: string;
  job_date: Date;
  order_type: string;
  confirm_date: Date;
  job_no: string;
  doc_ref: string;
  canceled: string;
  cancel_date: Date;
  invoiced: string;
  invoice_date: Date;
  created_at?: Date;
  created_by?: string;
  updated_at?: Date;
  updated_by?: string;
}
export interface IToOrderDetail {
  company_code: string;
  prin_code: string;
  job_no: string;
  cust_code: string;
  order_no: string;
  serial_no: number;
  prod_code: string;
  qty_puom: number;
  p_uom: string;
  qty_luom: number;
  quantity: number;
  doc_ref: string;
  lot_no: string;
  po_no: string;
  imp_job_no: string;
  manu_code: string;
  container_no: string;
  production_from: Date;
  production_to: Date;
  expiry_from: Date;
  expiry_to: Date;
  unit_price: number;
  site_code: string;
  loc_code_from: string;
  loc_code_to: string;
  picked: string;
  confirmed: string;
  confirmed_date: Date;
  l_uom: string;
  uppp: number;
  selected: string;
  aisle_from: string;
  aisle_to: string;
  height_from: string;
  height_to: string;
  column_from: string;
  column_to: string;
  gate_no: string;
  sales_rate: number;
  exp_container_no: string;
  exp_container_size: number;
  exp_container_type: string;
  exp_container_sealno: string;
  moc1: string;
  moc2: string;
  order_serial: number;
  origin_country: string;
  bal_pack_qty: number;
  multi_series: string;
  prod_attrib_code: string;
  prod_grade1: string;
  prod_grade2: string;
  tx_identity_number: string;
  ref_txn_code: string;
  ref_txn_slno: number;
  so_txn_code: string;
  inbound_done: string;
  ref_txn_doc: string;
  supp_code: string;
  supp_reference: string;
  orig_prod_code: string;
  salesman_code: string;
  hs_code: string;
  batch_no: string;
  act_order_qty: number;
  bal_order_qty: number;
  minperiod_exppick: number;
  ignore_minexp_period: string;
  stock_owner: string;
  ind_code: string;
  git_no: string;
  priority: string;
 qty_string: string;
}
export interface IPickwave {
  wave_name: string;
  wave_code: number;
  company_code: string;
  col_name: string;
  indicator: string;
  seq_order: string;
  updated_at: Date;
  updated_by: string;
  created_by: string;
  created_at: Date;
}